udp_server/pyproject.toml
[tool.poetry]
name = "udp-server"
version = "0.1.0"
description = ""
authors = ["RafalNogueiraXD <rafa10_nogueira@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
taskipy = "^1.14.0"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]
start = "python -m udp_server.main"
test = "pytest"
lint = "ruff check ."
format = "ruff format ."